refactor: move iqf to a sperate file
